src/assistant/requirements.txt
joblib
scikit-learn==1.6.1
groq
python-dotenv